+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Disable use of Geo API by default and added --geo-api option to enable it.
 - Added optional timezone configuration (needed for Docker).
 - Added packages needed when script is run inside Docker with Debian/Ubuntu.
+- Create /etc/localtime with a default timezone if it does not already exist.
 
 ## 1.1.1